fixing all sonar bugs

+++ src/main/java/code/components/render/subcomponents/ShieldRenderer.java
@@ -17,6 +17,6 @@ public class ShieldRenderer extends Renderer {
         NEntity link = ((LinkComponent) entity.getComponent(ComponentType.LINK)).getLink();
         ShieldItemComponent shc = (ShieldItemComponent) link.getComponent(ComponentType.ITEM);
 
-        data.setAlpha((shc.getCapacity() / shc.getMaxCapacity()) / ALPHA_REDUCER + ALPHA_OFFSET);
+        data.setAlpha(((double) shc.getCapacity() / shc.getMaxCapacity()) / ALPHA_REDUCER + ALPHA_OFFSET);
     }
 }
